refactor: create a custom hook to intergrate with nextauth and work with row leval security
